requirements.txt
django-redis-sessions-0.4.0